harness_requirements.txt
# SochDB Comprehensive Test Harness Requirements (v2.0 - Real LLM Mode)

# Core dependencies
numpy>=1.24.0
python-dotenv>=1.0.0

# Azure OpenAI integration (REQUIRED for real LLM mode)
openai>=1.12.0

# SochDB SDK (relative path)
# Install with: pip install -e ../sochdb-python-sdk

# Optional: For enhanced reporting
tabulate>=0.9.0
pandas>=2.0.0

# Required environment variables in .env:
# - AZURE_OPENAI_API_KEY
# - AZURE_OPENAI_ENDPOINT  
# - AZURE_OPENAI_EMBEDDING_DEPLOYMENT
# - AZURE_OPENAI_CHAT_DEPLOYMENT
# - AZURE_OPENAI_API_VERSION
